Collapse MemberPart into PartialMember
